+++ src/types/sdk/ClientConfiguration.ts
@@ -1,4 +1,5 @@
 export interface ClientConfiguration {
   bearerToken: string;
   baseUrl?: string;
+  stelaBaseUrl?: string;
 }