+++ src/types/stela/PaginatedItems.ts
@@ -1,0 +1,8 @@
+export interface PaginatedItems<T> {
+  items: T[];
+  pagination: {
+    nextCursor: string,
+    nextPage: string,
+    totalPages: number,
+  }
+}